@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_app/debug/log/log.dart';
 import 'package:flutter_app/src/ad/ad_manager.dart';
+import 'package:flutter_app/src/connector/moodle_webapi_connector.dart';
 import 'package:flutter_app/src/file/my_downloader.dart';
+import 'package:flutter_app/src/model/moodle_webapi/moodle_profile_entity.dart';
 import 'package:flutter_app/src/notifications/notifications.dart';
 import 'package:flutter_app/src/store/model.dart';
 import 'package:flutter_app/src/util/analytics_utils.dart';
@@ -24,15 +26,19 @@ class MainController extends GetxController {
   RxInt currentIndex = 0.obs;
   RxList<Widget> pageList = <Widget>[].obs;
 
+  var isProfileLoading = false.obs;
+  Rxn<MoodleProfileEntity> profile = Rxn();
+
   @override
   Future<void> onInit() async {
     super.onInit();
     await appInit();
+    await getMoodleProfile();
   }
 
   Future<void> appInit() async {
     var context = Get.context;
-    if(context == null) {
+    if (context == null) {
       throw Exception("BuildContext is null");
     }
 
@@ -40,12 +46,12 @@ class MainController extends GetxController {
     bool catchError = await Model.instance.getInstance(); //一定要先getInstance()不然無法取得資料
     try {
       if (!(await Model.instance.getAgreeContributor())) {
-        RouteUtils.toAgreePrivacyPolicyScreen();
-        Get.delete<MainController>();
+        await Get.delete<MainController>();
+        await RouteUtils.toAgreePrivacyPolicyScreen();
         return;
       }
       await RemoteConfigUtils.init();
-      if(context.mounted) {
+      if (context.mounted) {
         await LanguageUtils.init(context);
       }
       APPVersion.initAndCheck().then((needUpdate) {
@@ -59,11 +65,11 @@ class MainController extends GetxController {
       await Notifications.instance.init();
 
       pageList.addAll([
-        const CourseTablePage(),
-        const SubSystemPage(),
-        const CalendarPage(),
-        const ScoreViewerPage(),
-        const OtherPage()
+        CourseTablePage(),
+        SubSystemPage(),
+        CalendarPage(),
+        ScoreViewerPage(),
+        OtherPage()
       ]);
       Get.forceAppUpdate();
     } catch (e, stack) {
@@ -72,10 +78,25 @@ class MainController extends GetxController {
     if (catchError) {
       await Future.delayed(const Duration(milliseconds: 500));
       ErrorDialogParameter parameter =
-      ErrorDialogParameter(desc: R.current.loadDataFail);
+          ErrorDialogParameter(desc: R.current.loadDataFail);
       parameter.btnCancelOnPress = null;
       parameter.btnOkText = R.current.sure;
       await ErrorDialog(parameter).show();
+    }
+  }
+
+  Future<void> getMoodleProfile() async {
+    try {
+      if(Model.instance.getAccount().isEmpty || Model.instance.getPassword().isEmpty) {
+        await RouteUtils.toLoginScreen();
+      }
+
+      isProfileLoading.value = true;
+      profile.value = await MoodleWebApiConnector.getProfile();
+    } catch (e) {
+      Log.e(e);
+    } finally {
+      isProfileLoading.value = false;
     }
   }
 
