@@ -4,6 +4,7 @@ import 'package:flutter_app/src/config/app_colors.dart';
 import 'package:flutter_app/src/config/app_link.dart';
 import 'package:flutter_app/src/connector/core/connector.dart';
 import 'package:flutter_app/src/connector/core/connector_parameter.dart';
+import 'package:flutter_app/ui/components/custom_appbar.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
 
@@ -13,9 +14,7 @@ class PrivacyPolicyPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: Text(R.current.PrivacyPolicy),
-      ),
+      appBar: baseAppbar(title: R.current.PrivacyPolicy),
       body: FutureBuilder<String>(
         future: Connector.getDataByGet(
             ConnectorParameter(AppLink.privacyPolicyUrl)),
