@@ -8,6 +8,7 @@ import 'package:flutter_app/src/store/model.dart';
 import 'package:flutter_app/src/task/ntust/ntust_sub_system_task.dart';
 import 'package:flutter_app/src/task/task_flow.dart';
 import 'package:flutter_app/src/util/route_utils.dart';
+import 'package:flutter_app/ui/components/custom_appbar.dart';
 import 'package:flutter_app/ui/components/page/error_page.dart';
 import 'package:flutter_app/ui/pages/password/webmail_password_dialog.dart';
 import 'package:flutter_staggered_animations/flutter_staggered_animations.dart';
@@ -34,9 +35,7 @@ class _SubSystemPageState extends State<SubSystemPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: Text(R.current.informationSystem),
-      ),
+      appBar: mainAppbar(title: R.current.informationSystem),
       body: FutureBuilder<List<APTreeJson>>(
         future: initTask(),
         builder:
